Add doom emacs :)
Originally I didn't really want to use this as it is a bit heavy. But I
am using it as a sort of alternative to trying to hack neovim to become
a full blown IDE when I just want it as a quick editor. So going
forward, let's use both!