Move zsh back into .config, remove .profile explicit loader
Maybe i do like it more this way (.config/zsh)

Also removed explicit loader, it's not that bad to have it this way, I
think. Experimented with .zshenv but didn't really like that approach.
If we want additional customisations for iteration/testing on
interactive shell exports, we can just use aliasrc for it. Shouldn't
need to modify the non-interactive environment like that too much.